requirements.txt
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.1
